# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Asset Management System API (ISO 27001)
  version: 1.0.0
  description: RESTful API for IT Asset & ISMS Management with ISO 27001 compliance
  contact:
    name: API Support
servers:
  - url: http://localhost:3000/api/v1
    description: Development
paths:
...
  /admin/users:
    get:
      summary: List all users
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    post:
      summary: Create user (admin only)
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInput'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /admin/users/{id}:
    get:
      summary: Get user by ID
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    put:
      summary: Update user
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserInput'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    delete:
      summary: Delete user (soft delete / archive)
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
...
  /admin/groups/{id}/users:
    put:
      summary: Assign users to group
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
...
  /controls/tests:
    post:
      summary: Create a control implementation test with generic evidence links
      tags: [Controls]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ControlTestInput'
      responses:
        '201': { description: Control test created }
...
  /risks:
    get:
      summary: List risks (paginated, filterable)
      tags: [Risks]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'

  # ==================== Controls ====================
...
  /treatments:
    get:
      summary: List risk treatments
      tags: [Treatments]
      security: [{ bearerAuth: [] }]

  # ==================== Risk Methods ====================
  /methods:
    get:
      summary: List risk methods
      tags: [Methods]
      security: [{ bearerAuth: [] }]

  # ==================== Integrations ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ControlTestInput:
      type: object
      required: [controlImplementationId, testType, testedBy, result]
      properties:
        controlImplementationId: { type: string, format: uuid }
        testType: { type: string }
        testMethod: { type: string }
        testedBy: { type: string }
        testedAt: { type: string, format: date-time }
        result: { type: string }
        effectivenessRating: { type: integer, minimum: 0, maximum: 100 }
        findings: { type: string }
        evidenceRequired: { type: boolean }
        nextTestDate: { type: string, format: date-time }
        evidenceLinks:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceLinkInput'
    EvidenceLinkInput:
      type: object
      required: [evidenceId]
      properties:
        evidenceId: { type: string, format: uuid }
        relationType: { type: string, default: supports }
...
    RegisterInput:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 12 }
        firstName: { type: string }
        lastName: { type: string }
        phoneNumber: { type: string }
        organizationUnitId: { type: string }
    UpdateUserInput:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
        phoneNumber: { type: string }
        organizationUnitId: { type: string }
        isActive: { type: boolean }
...
    UserResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        displayId: { type: string }
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        phoneNumber: { type: string, nullable: true }
        isActive: { type: boolean }
        roles:
          type: array
          items: { type: string }
        language: { type: string, nullable: true }
        darkMode: { type: boolean, nullable: true }
        createdAt: { type: string, format: date-time }
...
    PaginatedResponse:
      type: object
      properties:
        data: { type: array, items: {} }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        totalPages: { type: integer }

# ==================== Phase 8: API, Betrieb und Nachweise ====================

  # ==================== Health & System (Phase 8) ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ==================== Phase 8 Schemas ====================
...